+++ packages/terra-search-field/src/terra-dev-site/doc/example/SearchFieldDisableAutoSearch.jsx
@@ -1,20 +1,12 @@
-import React, { useState } from 'react';
+import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import SearchField from 'terra-search-field/lib/SearchField';
 
-const INVALID_MESSAGE = 'The default minimum search length is 2.';
 
-const SearchFieldDisableAutoSearch = () => {
-  const [searchText, setSearchText] = useState('');
-
-  const message = searchText.length >= 2 ? `Search text: ${searchText}` : INVALID_MESSAGE;
-
-  return (
-    <div>
-      <p>{message}</p>
-      <SearchField disableAutoSearch onSearch={setSearchText} onInvalidSearch={setSearchText} />
-    </div>
-  );
-};
+const SearchFieldDisableAutoSearch = () => (
+  <div>
+    <SearchField disableAutoSearch />
+  </div>
+);
 
 export default SearchFieldDisableAutoSearch;