+++ packages/terra-search-field/src/terra-dev-site/doc/example/SearchFieldDisabled.jsx
@@ -1,20 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import SearchField from 'terra-search-field/lib/SearchField';
 
-const INVALID_MESSAGE = 'The default minimum search length is 2.';
-
-const SearchFieldDisabled = () => {
-  const [searchText, setSearchText] = useState('');
-
-  const message = searchText.length >= 2 ? `Search text: ${searchText}` : INVALID_MESSAGE;
-
-  return (
-    <div>
-      <p>{message}</p>
-      <SearchField isDisabled onSearch={setSearchText} onInvalidSearch={setSearchText} />
-    </div>
-  );
-};
+const SearchFieldDisabled = () => (
+  <SearchField isDisabled />
+);
 
 export default SearchFieldDisabled;