+++ packages/terra-avatar/CHANGELOG.md
@@ -3,6 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Fixed issue for avatar to allow non-square images by sending `Fit` as `cover`.
+
+### Changed
+* updated package.json test scripts
+
+2.23.0 - (July 30, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 2.22.0 - (July 24, 2019)
 ------------------